third_party/WebKit/LayoutTests/fast/css/css-font-face-src-value-uri.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
  test(function() {
    var cssText = '@font-face { src: url("font.woff") format("woff"), local("font2"); }';
    var doc = document.implementation.createHTMLDocument("");
    var styleElement = document.createElement("style");
    styleElement.textContent = cssText;
    doc.body.appendChild(styleElement);
    var cssRules = styleElement.sheet.cssRules;
    assert_equals(cssRules.length, 1);
    var compareCSSText = cssRules[0].cssText;
    assert_equals(compareCSSText, cssText);
  }, 'Testing serialization and round trip of URLs in CSSFontFaceSrcValue');
</script>
